Refactor the list command into its own file. Also update the description to be more accurate.
