+++ internal/boxcli/global.go
@@ -60,35 +60,6 @@ func addCommandAndHideConfigFlag(parent, child *cobra.Command) {
 	_ = child.Flags().MarkHidden("config")
 }
 
-type listCmdFlags struct {
-	config configFlags
-}
-
-func listCmd() *cobra.Command {
-	flags := listCmdFlags{}
-	cmd := &cobra.Command{
-		Use:     "list",
-		Aliases: []string{"ls"},
-		Short:   "List global packages",
-		PreRunE: ensureNixInstalled,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			box, err := devbox.Open(&devopt.Opts{
-				Dir:    flags.config.path,
-				Stderr: cmd.ErrOrStderr(),
-			})
-			if err != nil {
-				return errors.WithStack(err)
-			}
-			for _, p := range box.AllPackageNamesIncludingRemovedTriggerPackages() {
-				fmt.Fprintf(cmd.OutOrStdout(), "* %s\n", p)
-			}
-			return nil
-		},
-	}
-	flags.config.register(cmd)
-	return cmd
-}
-
 var globalConfigPath string
 
 func ensureGlobalConfig() (string, error) {
